optimize GPIO expander + external decoder fix
- external decoders sink callback had infinite loop when output buffer was full and would never empty
- race condition when playback stopped/restarted while waiting for output buffer to empty

+++ components/services/gpio_exp.h
@@ -16,7 +16,7 @@ struct gpio_exp_s;
 
 typedef struct {
 	char model[32];
-	uint8_t intr;
+	int intr;
 	uint8_t count;
 	uint32_t base;
 	struct gpio_exp_phy_s {
